added preloader.js
