+++ src/index.html
@@ -8,12 +8,13 @@
   </head>
 
   <body>
-    <div class="preloader" id="PreLoaderspiner"
-    style="display: none;">
+    <div class="preloader-con">
+    <div class="preloader" id="PreLoaderspiner">
       <div class="dach uno"></div>
       <div class="dach dos"></div>
       <div class="dach tres"></div>
       <div class="dach cuatro"></div>
+    </div>
     </div>
     <div class="form-container">
       <form class="search-form">
