+++ src/sass/components/preloader.scss
@@ -1,3 +1,9 @@
+.preloader-con {
+  position: absolute;
+  background-color: gray;
+  width: 100%;
+  height: 100%;
+}
 .preloader {
   position: absolute;
   z-index: 1000000000000000000;
